Tidy up intent handlers
